+++ client/src/screens/hub/HubMainScreen/models/TraceList.js
@@ -3,8 +3,6 @@ import {
   deepEqual,
   getObjectValueByPath,
   arraysIntersection,
-  formatValue,
-  flattenObject,
 } from '../../../../utils';
 import { COLORS } from '../../../../constants/colors';
 import { STROKES } from '../../../../constants/strokes';
@@ -143,7 +141,14 @@ export default class TraceList {
     }
   };
 
-  addSeries = (run, metric = null, trace = null, aggregate = false, seed) => {
+  addSeries = (
+    run,
+    metric = null,
+    trace = null,
+    alignBy = 'step',
+    aggregate = false,
+    seed,
+  ) => {
     let subGroup = this.groups;
     this.groupingFields.forEach((g) => {
       const groupVal = this.getRunParam(g, run, metric, trace);
@@ -310,10 +315,129 @@ export default class TraceList {
 
     // Add series to trace
     const seriesModel = new Series(run, metric, trace);
-    traceModel.addSeries(seriesModel, aggregate);
+
+    let alignment = alignBy;
+
+    if (alignBy === 'epoch' && this.grouping.chart.includes('context.subset')) {
+      alignment = 'step';
+    }
+    traceModel.addSeries(seriesModel, aggregate, alignment);
+    this.setAxisValues(alignBy, aggregate);
   };
 
   getChartsNumber = () => {
     return this.groupingConfigMap.charts.length;
+  };
+
+  setAxisValues = (alignBy, aggregate) => {
+    switch (alignBy) {
+      case 'step':
+        this.traces.forEach((traceModel) => {
+          traceModel.series.forEach((series) => {
+            const { trace } = series;
+            trace.axisValues = [];
+            trace.data.forEach((point) => {
+              trace.axisValues.push(point[1]);
+            });
+          });
+        });
+        break;
+      case 'epoch':
+        let epochSteps = {};
+        this.traces.forEach((traceModel) => {
+          traceModel.series.forEach((series) => {
+            const { trace } = series;
+            trace.data.forEach((point) => {
+              const epoch = point[2];
+              if (!epochSteps.hasOwnProperty(epoch)) {
+                epochSteps[epoch] = [];
+              }
+            });
+          });
+        });
+
+        for (let epoch in epochSteps) {
+          this.traces.forEach((traceModel) => {
+            traceModel.series.forEach((series) => {
+              const { trace } = series;
+              const stepsInEpoch = trace.data
+                .filter((point) => `${point[2]}` === epoch)
+                .map((point) => point[1]);
+              if (stepsInEpoch.length > epochSteps[epoch].length) {
+                if (epoch !== 'null' && +epoch > 0) {
+                  const prevEpoch = +epoch - 1;
+                  if (epochSteps.hasOwnProperty(prevEpoch)) {
+                    const diff = stepsInEpoch[1] - stepsInEpoch[0];
+                    const prevLastValue =
+                      epochSteps[prevEpoch][epochSteps[prevEpoch].length - 1];
+                    epochSteps[epoch] = stepsInEpoch.map(
+                      (step, i) => prevLastValue + (i + 1) * diff,
+                    );
+                  } else {
+                    epochSteps[epoch] = stepsInEpoch;
+                  }
+                } else {
+                  epochSteps[epoch] = stepsInEpoch;
+                }
+              }
+            });
+          });
+        }
+
+        this.epochSteps = epochSteps;
+
+        this.traces.forEach((traceModel) => {
+          traceModel.series.forEach((series) => {
+            const { trace } = series;
+            trace.axisValues = [];
+            for (let epoch in epochSteps) {
+              const stepsInEpoch = trace.data
+                .filter((point) => `${point[2]}` === epoch)
+                .map((point) => point[1]);
+              if (stepsInEpoch.length > 0) {
+                trace.axisValues =
+                  epoch === 'null'
+                    ? stepsInEpoch
+                    : trace.axisValues.concat(
+                      epochSteps[epoch].slice(
+                        epochSteps[epoch].length - stepsInEpoch.length,
+                      ),
+                    );
+              }
+            }
+          });
+        });
+        // Apply alignemnt to aggregated data
+        // TODO: improve overall aggregation logic (use it in single place)
+        if (aggregate) {
+          this.traces.forEach((traceModel) => {
+            const valuesByStep = {};
+            traceModel.series.forEach((series) => {
+              const { trace } = series;
+              trace.data.forEach((point, index) => {
+                const step = trace.axisValues[index];
+                if (!valuesByStep.hasOwnProperty(step)) {
+                  valuesByStep[step] = [];
+                }
+                let value = point[0];
+                valuesByStep[step].push(value);
+              });
+            });
+            traceModel.aggregation.max.trace.data = Object.keys(
+              valuesByStep,
+            ).map((step) => [_.max(valuesByStep[step]), +step]);
+            traceModel.aggregation.min.trace.data = Object.keys(
+              valuesByStep,
+            ).map((step) => [_.min(valuesByStep[step]), +step]);
+            traceModel.aggregation.avg.trace.data = Object.keys(
+              valuesByStep,
+            ).map((step) => [
+              _.sum(valuesByStep[step]) / valuesByStep[step].length,
+              +step,
+            ]);
+          });
+        }
+        break;
+    }
   };
 }