+++ client/src/ui/components/RangeSlider/RangeSlider.js
@@ -2,6 +2,8 @@ import './RangeSlider.less';
 
 import React from 'react';
 import PropTypes from 'prop-types';
+import Slider from 'react-rangeslider';
+import 'react-rangeslider/lib/index.css';
 
 import { classNames } from '../../utils';
 
@@ -12,15 +14,17 @@ class RangeSlider extends React.Component {
     this.state = {
       value: this.props.value,
     };
+    this.value = this.props.value;
   }
 
-  handleChange = (evt) => {
-    let newValue = evt.target.value;
+  handleChange = (value) => {
+    this.value = value;
+    this.setState({ value });
+  };
 
-    this.setState({ value: newValue });
-
+  handleChangeComplete = () => {
     if (this.props.onChange) {
-      this.props.onChange(newValue);
+      this.props.onChange(this.value);
     }
   };
 
@@ -30,15 +34,32 @@ class RangeSlider extends React.Component {
       [this.props.className]: !!this.props.className,
     });
 
+    let ticks = {};
+    if (!!this.props.ticks) {
+      ticks = this.props.ticks;
+    } else {
+      Array(6)
+        .fill(0)
+        .forEach((_, i) => {
+          const value = Math.round(
+            ((this.props.max - this.props.min) / 5) * i + this.props.min,
+          );
+          ticks[value] = value;
+        });
+    }
+
     return (
       <div className='RangeSlider__wrapper'>
-        <input
+        <Slider
           className={className}
-          type='range'
           min={this.props.min}
           max={this.props.max}
           value={this.state.value}
-          onChange={(evt) => this.handleChange(evt)}
+          handleLabel={`${this.state.value}`}
+          labels={ticks}
+          onChange={this.handleChange}
+          onChangeComplete={this.handleChangeComplete}
+          tooltip={false}
         />
       </div>
     );
@@ -49,12 +70,14 @@ RangeSlider.defaultProps = {
   min: 0,
   max: 100,
   value: 50,
+  ticks: null,
 };
 
 RangeSlider.propTypes = {
   min: PropTypes.number,
   max: PropTypes.number,
   value: PropTypes.number,
+  ticks: PropTypes.object,
   onChange: PropTypes.func,
 };